Aula 5

+++ Aula5/AppTarefas/js/script.js
@@ -1,20 +1,32 @@
 function AdicionaTarefa(){
     let minhaTarefa = $("#tarefa").val();
-    $("#lista-tarefas").append("<li>" + minhaTarefa + " <button type='button' onclick='ExcluirTarefa(this);'>Excluir</button> </li>");
-    $("#tarefa").val("");
     
-    SalvarTarefas();
+    if(minhaTarefa==""){
+        alert("Tarefa vazia, não pode ser adicionada");
+    }else{
+
+        $("#lista-tarefas").append("<li>" + minhaTarefa + " <button type='button' onclick='ExcluirTarefa(this);'>Excluir</button> </li>");
+        $("#tarefa").val("");
+        SalvarTarefas();
+    }
 }
 
 function ExcluirTarefa(botao){
     console.dir(botao);
     $(botao).parent().remove();
-    localStorage.tarefasSalvas = $("#lista-tarefas").html();
+    localStorage.listaDeTarefas = $("#lista-tarefas").html();
 }
 
 function ExcluirTodas(){
-    $("#lista-tarefas").html("");
-    localStorage.clear();
+    let resposta = confirm("Continuar com a exclusão?");
+    
+    if(resposta == true){
+        $("#lista-tarefas").html("");
+        localStorage.clear();
+        alert("Tudo apagado.");
+    }else{
+        alert("Exclusão cancelada.");
+    }
     /*
     https://speakerdeck.com/gutoffline/revisitando-local-storage
     */
